show location history by different facets

+++ src/main/java/co/uk/ticklethepanda/location/history/cartograph/heatmap/HeatmapColourPicker.java
@@ -7,45 +7,45 @@ import java.awt.*;
  */
 public interface HeatmapColourPicker {
 
-    default double normaliseValue(double unscaled, double largestNumber) {
-        return 1.0 - Math.log(unscaled) / Math.log(largestNumber);
+    default float normaliseValue(float unscaled, float largestNumber) {
+        return 1.0f - (float) Math.log(unscaled) / (float) Math.log(largestNumber);
     }
 
-    default Color pickColor(double unscaled, double largestNumber) {
+    default Color pickColor(float unscaled, float largestNumber) {
         return pickColor(normaliseValue(unscaled, largestNumber));
     }
 
-    Color pickColor(double percentage);
+    Color pickColor(float percentage);
 
     class Greyscale implements HeatmapColourPicker {
 
-        public static final double MAX_BRIGHTNESS = 0.2;
+        public static final float MAX_BRIGHTNESS = 0.2f;
 
-        private final double maximumBrightness;
+        private final float maximumBrightness;
 
         public Greyscale() {
             this(MAX_BRIGHTNESS);
         }
 
-        public Greyscale(double maximumBrightness) {
+        public Greyscale(float maximumBrightness) {
             this.maximumBrightness = maximumBrightness;
         }
 
         @Override
-        public Color pickColor(double percentage) {
-            double brightness = (1.0 - maximumBrightness) * percentage;
-            Color color = Color.getHSBColor(0f, 0f, (float) brightness);
+        public Color pickColor(float percentage) {
+            float brightness = (1.0f - maximumBrightness) * percentage;
+            Color color = Color.getHSBColor(0f, 0f, brightness);
             return color;
         }
     }
 
     class Monotone implements HeatmapColourPicker {
 
-        public static final double MIN_BRIGHTNESS = 0.1;
-        public static final double MAX_BRIGHTNESS = 0.9;
+        public static final float MIN_BRIGHTNESS = 0.1f;
+        public static final float MAX_BRIGHTNESS = 0.9f;
 
-        private final double minBrightness;
-        private final double maxBrightness;
+        private final float minBrightness;
+        private final float maxBrightness;
 
         private final Color color;
 
@@ -53,15 +53,15 @@ public interface HeatmapColourPicker {
             this(color, MIN_BRIGHTNESS, MAX_BRIGHTNESS);
         }
 
-        public Monotone(Color color, double minBrightness, double maxBrightness) {
+        public Monotone(Color color, float minBrightness, float maxBrightness) {
             this.color = color;
             this.minBrightness = minBrightness;
             this.maxBrightness = maxBrightness;
         }
 
         @Override
-        public Color pickColor(double percentage) {
-            double brightness = (maxBrightness - minBrightness) * percentage + minBrightness;
+        public Color pickColor(float percentage) {
+            float brightness = (maxBrightness - minBrightness) * percentage + minBrightness;
 
             float[] hsb = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null);
 
@@ -71,8 +71,8 @@ public interface HeatmapColourPicker {
     }
 
     class Hue implements HeatmapColourPicker {
-        public Color pickColor(double percentage) {
-            double hue = percentage;
+        public Color pickColor(float percentage) {
+            float hue = percentage;
             hue = Math.min(hue, 1f) * 0.8f;
             return Color.getHSBColor(
                     (float) hue, 1f, 1f);
